aboutme.html
Hello! I've created this website through the program "Girls Who Code." I was curious to know what on earth I'd be doing over the summer, and now I'm blessed with coming here. As of now I have been working on "scratches," and learning how to create websites. These tasks seem to be the basics of what I need to know through coding. In this website, I'll be sharing my coding experience and what I feel is my place in Girls Who Code.